requirements.txt
# FastAPI & Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# LangChain & LLM Integration (Pinned for compatibility)
langchain==1.0.2
langchain-core==1.2.18
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-google-genai==4.2.1
langchain-openai==1.0.2
# Note: langchain-chroma is optional; we use fallback import

# Graph & Workflow
langgraph==1.0.1
langgraph-checkpoint>=1.0.0

# Vector Database & Knowledge Graph
chromadb>=0.5.0
networkx>=3.0

# Utilities
python-dotenv>=0.19.0
requests>=2.28.0
httpx>=0.24.0
faker>=18.0.0

# Testing & Development
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=23.0.0
flake8>=6.0.0
